+++ lib/file.php
@@ -107,9 +107,15 @@ function get_filename($page)
 // Put a data(wiki text) into a physical file(diff, backup, text)
 function page_write($page, $postdata, $notimestamp = FALSE)
 {
-	global $autoalias, $aliaspage;
+	global $autoalias, $aliaspage, $edit_auth;
 
 	if (PKWK_READONLY) return; // Do nothing
+
+	// Defense in depth: block writes when edit_auth denies the page
+	if ($edit_auth && ! is_page_writable($page)) {
+		ensure_page_writable($page);
+		return;
+	}
 
 	$postdata = make_str_rules($postdata);
 	$timestamp_to_keep = null;
@@ -319,7 +325,7 @@ function file_head($file, $count = 1, $lock = TRUE, $buffer = 8192)
 
 	$fp = @fopen($file, 'r');
 	if ($fp === FALSE) return FALSE;
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	if ($lock) flock($fp, LOCK_SH);
 	rewind($fp);
 	$index = 0;
@@ -386,7 +392,7 @@ function file_write($dir, $page, $str, $notimestamp = FALSE, $is_delete = FALSE)
 		htmlsc(basename($dir) . '/' . encode($page) . '.txt') .	
 		'<br />' . "\n" .
 		'Maybe permission is not writable or filename is too long');
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	flock($fp, LOCK_EX);
 	ftruncate($fp, 0);
 	rewind($fp);
@@ -403,7 +409,7 @@ function file_write($dir, $page, $str, $notimestamp = FALSE, $is_delete = FALSE)
 
 		// Command execution per update
 		if (defined('PKWK_UPDATE_EXEC') && PKWK_UPDATE_EXEC)
-			system(PKWK_UPDATE_EXEC . ' > /dev/null &');
+			pkwk_run_update_exec();
 
 	} else if ($dir == DIFF_DIR && $notify) {
 		if ($notify_diff_only) $str = preg_replace('/^[^-+].*\n/m', '', $str);
@@ -450,7 +456,7 @@ function add_recent($page, $recentpage, $subject = '', $limit = 0)
 		die_message('Cannot write page file ' .
 		htmlsc($recentpage) .
 		'<br />Maybe permission is not writable or filename is too long');
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	flock($fp, LOCK_EX);
 	rewind($fp);
 	fputs($fp, '#freeze'    . "\n");
@@ -485,7 +491,7 @@ function lastmodified_add($update = '', $remove = '')
 	pkwk_touch_file($file);
 	$fp = fopen($file, 'r+') or
 		die_message('Cannot open ' . 'CACHE_DIR/' . PKWK_MAXSHOW_CACHE);
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	flock($fp, LOCK_EX);
 
 	// Read (keep the order of the lines)
@@ -530,7 +536,7 @@ function lastmodified_add($update = '', $remove = '')
 	pkwk_touch_file($file);
 	$fp = fopen($file, 'r+') or
 		die_message('Cannot open ' . htmlsc($whatsnew));
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	flock($fp, LOCK_EX);
 
 	// Recreate
@@ -582,7 +588,7 @@ function put_lastmodified()
 	pkwk_touch_file($file);
 	$fp = fopen($file, 'r+') or
 		die_message('Cannot open' . 'CACHE_DIR/' . PKWK_MAXSHOW_CACHE);
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	flock($fp, LOCK_EX);
 	ftruncate($fp, 0);
 	rewind($fp);
@@ -596,7 +602,7 @@ function put_lastmodified()
 	pkwk_touch_file($file);
 	$fp = fopen($file, 'r+') or
 		die_message('Cannot open ' . htmlsc($whatsnew));
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	flock($fp, LOCK_EX);
 	ftruncate($fp, 0);
 	rewind($fp);
@@ -671,7 +677,7 @@ function autolink_pattern_write($filename, $autolink_pattern)
 
 	$fp = fopen($filename, 'w') or
 		die_message('Cannot open ' . $filename);
-	set_file_buffer($fp, 0);
+	pkwk_set_file_buffer($fp, 0);
 	flock($fp, LOCK_EX);
 	rewind($fp);
 	fputs($fp, $pattern   . "\n");
@@ -916,7 +922,7 @@ function get_readings()
 
 					$readings[$page] = $page;
 					foreach ($patterns as $no => $pattern)
-						$readings[$page] = mb_convert_kana(mb_ereg_replace($pattern,
+						$readings[$page] = mb_convert_kana(preg_replace('/' . $pattern . '/u',
 							$replacements[$no], $readings[$page]), 'aKCV');
 				}
 				break;